fix: share admin runtime core without detail loading
