+++ src/features/admin/legacy/script.ts
@@ -1,55 +1,81 @@
 import type { LegacyDomElement, LegacyHistoryEntry, LegacyModalOptions, LegacyPageState, LegacyRecord, LegacyRuntimeData } from "./runtime";
 import { reportSubmissionSchema } from "../data/admin-records";
-import { data, recordsFor } from "./runtime-data";
+import {
+  adminDateTime,
+  applyDemoAction as applyDemoActionCore,
+  applyReportDecision as applyReportDecisionCore,
+  autoRejectUnavailablePayout as autoRejectUnavailablePayoutCore,
+  badge,
+  completedPayoutQuests,
+  confirmedViolationCount,
+  currentAdminName,
+  disputeTypeLabel,
+  escapeActivityText,
+  fmt,
+  formatActivityTime,
+  ico,
+  penaltyOutcomeFor,
+  penaltyOutcomeLabel,
+  payoutEarningForQuest,
+  payoutFinancials,
+  payoutPreviousRecords,
+  payoutDecisionContext,
+  readActivityEvents,
+  redFlagExemptionFor,
+  recordActivity,
+  recordConfirmedViolation,
+  reportDateTime,
+  timeline,
+  toneClass,
+  userQuestRecords,
+  userReportsFor,
+  data,
+} from "./runtime-core";
+import { createOverlayRuntime } from "./overlay-runtime";
+import { setActiveNavigation as setActiveNavigationCore } from "./navigation-state";
+import { recordsFor } from "./runtime-data";
 
-export { data, disputeCases } from "./runtime-data";
+export {
+  addUserHistory,
+  adminDateTime,
+  badge,
+  bindChatAttachment,
+  chatMessage,
+  chatTimeLabel,
+  completedPayoutQuests,
+  confirmedViolationCount,
+  currentAdminName,
+  disputeTypeLabel,
+  escapeActivityText,
+  fmt,
+  ico,
+  penaltyOutcomeFor,
+  penaltyOutcomeLabel,
+  payoutEarningForQuest,
+  payoutFinancials,
+  payoutPreviousRecords,
+  payoutQuestId,
+  payoutTimestamp,
+  payoutDecisionContext,
+  redFlagExemptionFor,
+  recordActivity,
+  reportDateTime,
+  seedGeneratedActivity,
+  timeline,
+  toneClass,
+  userQuestRecords,
+  userReportsFor,
+  data,
+  disputeCases,
+} from "./runtime-core";
 
 type LegacyView = "home" | "disputes" | "quests" | "users" | "payouts" | "reports" | "policies" | "activity";
 type IconName = "home" | "scale" | "quest" | "users" | "wallet" | "settings" | "history" | "menu" | "search" | "filter" | "paperclip" | "check" | "user" | "flag";
-type Tone = "warning" | "danger" | "success" | "info" | "neutral" | "assigned" | "cancelled";
 type LegacyForm = HTMLFormElement & {
   elements: HTMLFormControlsCollection & Record<string, LegacyDomElement>;
 };
-type TimelineEntry = string | { title: string; detail?: string; time?: string; showDetails?: boolean };
-type TimelineOptions = { showDetails?: boolean };
-type PenaltyOutcome = { key?: string; label: string; status: string; tone: string; durationDays?: number | null };
-type PenaltyExemption = { remaining: number; label: string; field: string };
 type ConfirmActionOptions = LegacyModalOptions;
 
-const paths: Record<IconName, string> = {
-  home: '<path d="M3 11.5 12 4l9 7.5"/><path d="M5.5 10v10h13V10M9 20v-6h6v6"/>',
-  scale: '<path d="M12 3v18M5 7h14M5 7l-3 6h6L5 7Zm14 0-3 6h6l-3-6ZM8 21h8"/>',
-  quest:
-    '<rect x="5" y="4" width="14" height="17" rx="2"/><path d="M9 4V3h6v1M8 9h8M8 13h8M8 17h5"/>',
-  users:
-    '<path d="M16 21v-2a4 4 0 0 0-4-4H6a4 4 0 0 0-4 4v2M9 11a4 4 0 1 0 0-8 4 4 0 0 0 0 8M22 21v-2a4 4 0 0 0-3-3.8"/>',
-  wallet:
-    '<path d="M3 6h16a2 2 0 0 1 2 2v10a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V6Zm0 0 12-3v3"/><path d="M16 12h5v4h-5a2 2 0 0 1 0-4Z"/>',
-  settings:
-    '<path d="M4 21v-7M4 10V3M12 21v-9M12 8V3M20 21v-5M20 12V3M1 14h6M9 8h6M17 16h6"/>',
-  history: '<path d="M3 12a9 9 0 1 0 3-6.7L3 8M3 3v5h5M12 7v5l3 2"/>',
-  menu: '<path d="M4 7h16M4 12h16M4 17h16"/>',
-  search: '<circle cx="11" cy="11" r="7"/><path d="m20 20-4-4"/>',
-  filter: '<path d="M4 5h16M7 12h10M10 19h4"/>',
-  paperclip: '<path d="m21.4 11.6-8.5 8.5a6 6 0 0 1-8.5-8.5l9.2-9.2a4 4 0 0 1 5.7 5.7l-9.2 9.2a2 2 0 0 1-2.8-2.8l8.5-8.5"/>',
-  check: '<path d="m4 12 5 5L20 6"/>',
-  user: '<circle cx="12" cy="8" r="4"/><path d="M4 21a8 8 0 0 1 16 0"/>',
-  flag: '<path d="M5 21V4m0 0h12l-2 4 2 4H5"/>',
-};
-export const ico = (n: string): string =>
-  `<svg class="ui-icon" viewBox="0 0 24 24" aria-hidden="true">${paths[n as IconName] || paths.quest}</svg>`;
-const allowedTones: Tone[] = [
-  "warning",
-  "danger",
-  "success",
-  "info",
-  "neutral",
-  "assigned",
-  "cancelled",
-];
-export function toneClass(tone: string): Tone {
-  return allowedTones.includes(tone as Tone) ? tone as Tone : "neutral";
-}
 const navItems: Array<[LegacyView, IconName, string, string]> = [
   ["home", "home", "Overview", ""],
   ["quests", "quest", "Quests", ""],
@@ -83,30 +109,6 @@ export const state: LegacyPageState = { view: initialView, tab: "all", query: ""
 const mainElement = document.querySelector<LegacyDomElement>("main");
 if (!mainElement) throw new Error("Legacy admin main element is required");
 export const main: HTMLElement = mainElement;
-export const fmt = (n: number | null | undefined): string => new Intl.NumberFormat("en-US").format(Number(n || 0));
-const questStatusClass = (status: string): string => {
-    const slug = String(status ?? "")
-      .toLowerCase()
-      .replace(/[^a-z0-9]+/g, "-")
-      .replace(/^-|-$/g, "");
-    return [
-      "open",
-      "assigned",
-      "in-progress",
-      "submitted",
-      "change-pending",
-      "approved",
-      "disputed",
-      "completed",
-      "hidden",
-      "draft",
-      "cancelled",
-    ].includes(slug)
-      ? ` quest-status-${slug}`
-      : "";
-  };
-export const badge = (s: string, t: string): string =>
-    `<span class="badge ${toneClass(t)}${questStatusClass(s)}">${escapeActivityText(s)}</span>`;
 const nav = document.querySelector<LegacyDomElement>("#nav");
 if (!nav) throw new Error("Legacy admin navigation element is required");
 nav.innerHTML = navItems
@@ -263,223 +265,6 @@ export function activityList() {
       `<ul class="activity"><li><span class="avatar">${escapeActivityText(a[0])}</span><span><strong>${escapeActivityText(a[1])}</strong><p>${escapeActivityText(a[2])}</p><time>${escapeActivityText(a[3])}</time></span></li></ul>`,
   ) : ['<div class="empty"><h3>No activity recorded</h3><p>Administrative activity will appear here as actions are taken.</p></div>'];
 }
-export function escapeActivityText(value: unknown): string {
-  return String(value ?? "").replace(/[&<>"']/g, (character) => ({
-    "&": "&amp;",
-    "<": "&lt;",
-    ">": "&gt;",
-    '"': "&quot;",
-    "'": "&#39;",
-  })[character] || "");
-}
-export function chatTimeLabel(date = new Date()): string {
-  return `Today · ${date.toLocaleTimeString("en-GB", {
-    hour: "2-digit",
-    minute: "2-digit",
-  })}`;
-}
-export function reportDateTime(date = new Date()): string {
-  const datePart = date.toLocaleDateString("en-GB", {
-      day: "2-digit",
-      month: "short",
-      year: "numeric",
-      timeZone: "Asia/Bangkok",
-    }),
-    timePart = date.toLocaleTimeString("en-GB", {
-      hour: "2-digit",
-      minute: "2-digit",
-      hour12: false,
-      timeZone: "Asia/Bangkok",
-    });
-  return `${datePart} · ${timePart}`;
-}
-export function adminDateTime(date = new Date()): string {
-  return reportDateTime(date).replace(/\s+ICT$/, "");
-}
-export function currentAdminName(): string {
-  return "Nicha P.";
-}
-export function disputeTypeLabel(record: Pick<LegacyRecord, "disputeType">): string {
-  return record.disputeType || "Other";
-}
-export function chatMessage(sender: string, time: string, message: string, variant: string): string {
-  return `<article class="chat-message ${variant}"><div class="chat-message-meta"><strong>${escapeActivityText(sender)}</strong><time>${escapeActivityText(time)}</time></div><p>${escapeActivityText(message)}</p></article>`;
-}
-export function bindChatAttachment(form: HTMLFormElement): void {
-  const input = form.querySelector<LegacyDomElement>("[data-chat-attachment]"),
-    label = form.querySelector<LegacyDomElement>("[data-chat-attachment-name]");
-  if (!input || !label) return;
-  input.addEventListener("change", () => {
-    const file = input.files?.[0];
-    label.textContent = file?.name || "No file attached";
-    label.title = file?.name || "";
-  });
-}
-const activityStorageKey = "kuquest-admin-activity-v2";
-const activityDataVersionKey = "kuquest-admin-activity-version";
-const activityDataVersion = "2026-08-28-large-refresh-v2";
-try {
-  if (localStorage.getItem(activityDataVersionKey) !== activityDataVersion) {
-    localStorage.removeItem(activityStorageKey);
-    localStorage.setItem(activityDataVersionKey, activityDataVersion);
-  }
-} catch {
-  // Keep activity empty if browser storage is blocked.
-}
-function readActivityEvents() {
-  try {
-    const events = JSON.parse(localStorage.getItem(activityStorageKey) || "[]");
-    return Array.isArray(events) ? events : [];
-  } catch {
-    return [];
-  }
-}
-function activityTimestamp(value: unknown, fallback: number): number {
-  const timestamp = Date.parse(String(value ?? "").replace(" · ", " "));
-  return Number.isFinite(timestamp) ? timestamp : fallback;
-}
-function activityInitials(name: unknown, fallback: string): string {
-  const initials = String(name ?? "")
-    .trim()
-    .split(/\s+/)
-    .filter(Boolean)
-    .map((part) => part[0])
-    .join("")
-    .slice(0, 2)
-    .toUpperCase();
-  return initials || fallback;
-}
-export function seedGeneratedActivity(records: LegacyRuntimeData): void {
-  if (readActivityEvents().length) return;
-  const events: Array<{ actor: string; title: string; detail: string; timestamp: number }> = [];
-  const now = Date.now();
-  const add = (actor: string, title: string, detail: string, at: unknown, fallback: number): void => {
-    events.push({
-      actor,
-      title,
-      detail,
-      timestamp: activityTimestamp(at, fallback),
-    });
-  };
-
-  (records.reports || []).slice(0, 14).forEach((record, index) => {
-    const fallback = now - (index + 1) * 38 * 60 * 1000;
-    add(
-      activityInitials(record.reporterName, "US"),
-      "User report received",
-      `${record.id} · ${record.reporterName} reported ${record.reportedUserName}`,
-      record.reportedAt,
-      fallback,
-    );
-    if (record.status === "Closed") {
-      add(
-        "NP",
-        "Report resolved",
-        `${record.id} · ${record.decisionLabel || "Report closed"}`,
-        record.resolutionAt || record.closedAt,
-        fallback + 47 * 60 * 1000,
-      );
-    }
-  });
-
-  (records.disputes || []).slice(0, 12).forEach((record, index) => {
-    const fallback = now - (index + 2) * 17 * 60 * 60 * 1000;
-    add(
-      "SYS",
-      "Dispute opened",
-      `${record.id} · ${record.title}`,
-      record.disputeDate,
-      fallback,
-    );
-    if (record.status === "Closed") {
-      add(
-        "NP",
-        "Dispute resolved",
-        `${record.id} · ${record.resolution || "Resolution recorded"}`,
-        record.disputeDate,
-        fallback + 3 * 60 * 60 * 1000,
-      );
-    }
-  });
-
-  (records.payouts || []).slice(0, 12).forEach((record, index) => {
-    const fallback = now - (index + 1) * 21 * 60 * 60 * 1000;
-    add(
-      "SYS",
-      "Payout requested",
-      `${record.id} · ${record.title} · ฿${fmt(record.amount)}`,
-      record.requestedAt,
-      fallback,
-    );
-    if (record.status === "Processing" || record.status === "Completed") {
-      add(
-        "NP",
-        "Payout approved",
-        `${record.id} · ${record.title}`,
-        record.approvedAt,
-        fallback + 2 * 60 * 60 * 1000,
-      );
-    } else if (record.status === "Rejected") {
-      add(
-        "NP",
-        "Payout rejected",
-        `${record.id} · ${record.rejectionReason}`,
-        record.rejectedAt,
-        fallback + 2 * 60 * 60 * 1000,
-      );
-    }
-  });
-
-  (records.users || []).slice(0, 16).forEach((user, index) => {
-    (user.moderationHistory || []).slice(0, 2).forEach((entry, entryIndex) => {
-      add(
-        activityInitials(entry.by, entry.by === "System" ? "SYS" : "NP"),
-        entry.event || "Activity",
-        `${user.id} · ${user.title}${entry.reason || entry.note ? ` · ${entry.reason || entry.note}` : ""}`,
-        entry.at,
-        now - (index * 2 + entryIndex + 1) * 29 * 60 * 60 * 1000,
-      );
-    });
-  });
-
-  events.sort((first, second) => second.timestamp - first.timestamp);
-  try {
-    localStorage.setItem(activityStorageKey, JSON.stringify(events.slice(0, 40)));
-  } catch {
-    // Keep activity empty if browser storage is blocked.
-  }
-}
-export function recordActivity(title: string, detail: string, actor = "NP"): void {
-  const event = {
-    actor,
-    title: String(title),
-    detail: String(detail),
-    timestamp: Date.now(),
-  };
-  try {
-    localStorage.setItem(
-      activityStorageKey,
-      JSON.stringify([event, ...readActivityEvents()].slice(0, 40)),
-    );
-  } catch {
-    // Ignore activity writes if browser storage is blocked.
-  }
-}
-function formatActivityTime(timestamp: number): string {
-  const age = Math.max(0, Date.now() - Number(timestamp || 0));
-  const minutes = Math.floor(age / 60000);
-  if (minutes < 1) return "Just now";
-  if (minutes < 60) return `${minutes} minute${minutes === 1 ? "" : "s"} ago`;
-  const hours = Math.floor(minutes / 60);
-  if (hours < 24) return `${hours} hour${hours === 1 ? "" : "s"} ago`;
-  const days = Math.floor(hours / 24);
-  if (days < 7) return `${days} day${days === 1 ? "" : "s"} ago`;
-  return new Date(timestamp).toLocaleDateString("en-GB", {
-    day: "numeric",
-    month: "short",
-    year: "numeric",
-  });
-}
 export let renderResource = function renderResource(v: string): void {
   if (v === "policies") return renderPolicies();
   if (v === "activity") return renderActivity();
@@ -556,14 +341,7 @@ export function render() {
   setActiveNavigation(state.view);
 }
 export function setActiveNavigation(view: string): void {
-  document
-    .querySelectorAll<LegacyDomElement>("[data-view]")
-    .forEach((b) => {
-      const active = b.dataset.view === view;
-      b.classList.toggle("active", active);
-      if (active) b.setAttribute("aria-current", "page");
-      else b.removeAttribute("aria-current");
-    });
+  setActiveNavigationCore(document, view);
 }
 export let bind = function bind(): void {
   document.querySelectorAll<LegacyDomElement>(".filter").forEach((b) => {
@@ -650,197 +428,14 @@ if (!drawerElement || !scrimElement || !shellElement) throw new Error("Legacy ov
 export const drawer: HTMLElement = drawerElement;
 export const scrim: HTMLElement = scrimElement;
 export const shell: HTMLElement = shellElement;
-const focusableSelector: string = [
-    "a[href]",
-    "button:not([disabled])",
-    "input:not([disabled])",
-    "select:not([disabled])",
-    "textarea:not([disabled])",
-    '[tabindex]:not([tabindex="-1"])',
-  ].join(",");
-let drawerTrigger: Element | null = null,
-  drawerKeydown: ((event: KeyboardEvent) => void) | null = null,
-  activeCustomLayerClose: (() => void) | null = null;
-
-export function closeActiveLayer(): void {
-  activeCustomLayerClose?.();
-}
-
-function visibleFocusable(root: HTMLElement): HTMLElement[] {
-  return [...root.querySelectorAll<LegacyDomElement>(focusableSelector)].filter(
-    (element) => element.getClientRects().length && !element.closest("[hidden]"),
-  );
-}
-
-function trapFocus(event: KeyboardEvent, layer: HTMLElement): void {
-  if (event.key !== "Tab") return;
-  const focusable = visibleFocusable(layer);
-  if (!focusable.length) {
-    event.preventDefault();
-    layer.focus();
-    return;
-  }
-  const first = focusable[0],
-    last = focusable.at(-1);
-  if (event.shiftKey && document.activeElement === first) {
-    event.preventDefault();
-    last?.focus();
-  } else if (!event.shiftKey && document.activeElement === last) {
-    event.preventDefault();
-    first.focus();
-  }
-}
-
-export function showDrawerLayer(): void {
-  drawerTrigger = document.activeElement;
-  shell.inert = true;
-  drawer.inert = false;
-  scrim.hidden = false;
-  drawer.setAttribute("aria-hidden", "false");
-  drawerKeydown = (event) => trapFocus(event, drawer);
-  drawer.addEventListener("keydown", drawerKeydown);
-  requestAnimationFrame(() => {
-    drawer.classList.add("open");
-    const title = drawer.querySelector<LegacyDomElement>("h2")?.textContent?.trim();
-    drawer.setAttribute("aria-label", title ? `Record details: ${title}` : "Record details");
-    drawer.querySelector<LegacyDomElement>("#close")?.focus();
-  });
-}
-
-export function showModalLayer(layer: HTMLElement, options: LegacyModalOptions = {}): () => void {
-  activeCustomLayerClose?.();
-  const trigger = document.activeElement,
-    removeOnClose = options.removeOnClose !== false;
-  if (!layer.isConnected) document.body.append(layer);
-  const siblings = [...document.body.children]
-    .filter((element) => element !== layer)
-    .map((element) => ({ element: element as LegacyDomElement, inert: (element as LegacyDomElement).inert }));
-  siblings.forEach(({ element }) => (element.inert = true));
-  const onKeydown = (event: KeyboardEvent) => {
-    if (event.key === "Escape") {
-      event.preventDefault();
-      event.stopPropagation();
-      close();
-      return;
-    }
-    trapFocus(event, layer);
-  };
-  const close = () => {
-    layer.removeEventListener("keydown", onKeydown);
-    siblings.forEach(({ element, inert }) => (element.inert = inert));
-    if (removeOnClose) layer.remove();
-    options.onClose?.();
-    if (trigger instanceof HTMLElement && trigger.isConnected) trigger.focus();
-    if (activeCustomLayerClose === close) activeCustomLayerClose = null;
-  };
-  activeCustomLayerClose = close;
-  layer.addEventListener("keydown", onKeydown);
-  requestAnimationFrame(() => {
-    const preferred =
-      typeof options.initialFocus === "string"
-        ? layer.querySelector<LegacyDomElement>(options.initialFocus)
-        : options.initialFocus;
-    (preferred || visibleFocusable(layer)[0] || layer).focus();
-  });
-  return close;
-}
-function payoutDecisionContext(record: LegacyRecord): { heading: string; copy: string; next: string } {
-  const contexts = {
-    "Needs approval": {
-      heading: "Why your approval is needed",
-      copy: "This payout is ready for release, but it cannot move to the bank until an admin approves it.",
-      next: "Approve payout → status changes to Processing. Funds are not yet transferred.",
-    },
-    Processing: {
-      heading: "Transfer in progress",
-      copy: "The payout has been approved and is moving to the recipient’s bank. No action is needed unless the transfer fails.",
-      next: "The record will become Completed after the bank confirms the transfer.",
-    },
-    Completed: {
-      heading: "Transfer completed",
-      copy: "The recipient’s bank transfer completed successfully. This record is retained for audit.",
-      next: "No further admin action is available.",
-    },
-    Rejected: {
-      heading: "Payout rejected",
-      copy: "This payout was rejected before funds were released. Review the recorded reason before creating a new payout request.",
-      next: "No retry is available from this record.",
-    },
-    Failed: {
-      heading: "Transfer failed",
-      copy: "The payment provider could not complete this transfer.",
-      next: "Review the failure reason before creating a new payout request.",
-    },
-  };
-  return contexts[record.status as keyof typeof contexts] || contexts["Needs approval"];
-}
-function payoutPreviousRecords(record: LegacyRecord): LegacyRecord[] {
-  const currentTimestamp = payoutTimestamp(record);
-  return data.payouts
-    .filter((payout) => {
-      if (payout.title !== record.title || payout.id === record.id) return false;
-      const timestamp = payoutTimestamp(payout);
-      return timestamp < currentTimestamp || (!timestamp && !currentTimestamp);
-    })
-    .sort((first, second) => payoutTimestamp(second) - payoutTimestamp(first));
-}
-export function payoutTimestamp(record: LegacyRecord): number {
-  return Date.parse(String(record.requestedAt || "").replace(" · ", " ")) || 0;
-}
-export function payoutEarningForQuest(quest: LegacyRecord): number {
-  if (!quest.teamQuest) return Number(quest.amount || 0);
-  const workerCount = quest.teamParticipants?.length || Number(quest.teamSize) || 1;
-  return Math.round(Number(quest.amount || 0) / workerCount);
-}
-function payoutFinancials(record: LegacyRecord): { earned: number; committed: number; pending: number; available: number; remaining: number; previousPaidOut: number } {
-  const previousPayouts = payoutPreviousRecords(record),
-    processingReserved = previousPayouts
-      .filter((payout) => payout.status === "Processing")
-      .reduce((total, payout) => total + Number(payout.amount || 0), 0),
-    pendingReserved = previousPayouts
-      .filter((payout) => payout.status === "Needs approval")
-      .reduce((total, payout) => total + Number(payout.amount || 0), 0);
-  const previousPaidOut =
-      Number(record.previouslyPaidOut ??
-      previousPayouts
-        .filter((payout) => payout.status === "Completed")
-        .reduce((total, payout) => total + Number(payout.amount || 0), 0)),
-    earned = completedPayoutQuests(record).reduce(
-      (total, quest) => total + payoutEarningForQuest(quest),
-      0,
-    ),
-    balanceBeforeRequest = Math.max(
-      0,
-      earned - previousPaidOut - processingReserved - pendingReserved,
-    ),
-    available = balanceBeforeRequest,
-    remaining = ["Rejected", "Failed"].includes(record.status)
-      ? balanceBeforeRequest
-      : Math.max(0, balanceBeforeRequest - Number(record.amount || 0));
-  return {
-    earned,
-    committed: processingReserved + pendingReserved,
-    pending: pendingReserved,
-    available,
-    remaining,
-    previousPaidOut,
-  };
-}
+const overlayRuntime = createOverlayRuntime({ drawer, scrim, shell });
+export const closeActiveLayer = overlayRuntime.closeActiveLayer;
+export const closeDrawer = overlayRuntime.closeDrawer;
+export const showDrawerLayer = overlayRuntime.showDrawerLayer;
+export const showModalLayer = overlayRuntime.showModalLayer;
 function payoutSummarySection(record: LegacyRecord): string {
   const financials = payoutFinancials(record);
   return `<section class="section payout-summary"><h3>Payout summary</h3><div class="payout-summary-grid"><div><span>Available to withdraw</span><strong>฿${fmt(financials.available)}</strong></div><div><span>Payout amount</span><strong>฿${fmt(record.amount)}</strong></div><div><span>Remaining after payout</span><strong>฿${fmt(financials.remaining)}</strong></div><div><span>Previously paid out</span><strong>฿${fmt(financials.previousPaidOut)}</strong></div></div></section>`;
-}
-export function autoRejectUnavailablePayout(record: LegacyRecord): boolean {
-  if (record.status !== "Needs approval") return false;
-  const available = payoutFinancials(record).available;
-  if (available >= Number(record.amount || 0)) return false;
-  record.status = "Rejected";
-  record.tone = "danger";
-  record.rejectedAt = payoutDateTime();
-  record.rejectedBy = payoutAdminName();
-  record.rejectionReason = "Insufficient withdrawable balance.";
-  record.rejectionNote = "Automatically rejected before admin review because the available balance did not cover the request.";
-  return true;
 }
 function payoutTimingSection(record: LegacyRecord): string {
   const events: Array<[string, string]> = [["Requested", String(record.requestedAt || "Not recorded")]];
@@ -855,19 +450,6 @@ function payoutOutcomeSection(record: LegacyRecord): string {
   const reason = record.rejectionReason || record.failureReason;
   if (!reason || !["Rejected", "Failed"].includes(record.status)) return "";
   return `<section class="section payout-outcome"><h3>${record.status === "Failed" ? "Transfer failure reason" : "Rejection reason"}</h3><p>${escapeActivityText(reason)}</p>${record.rejectionNote ? `<p class="payout-admin-note"><strong>Admin note:</strong> ${escapeActivityText(record.rejectionNote)}</p>` : ""}</section>`;
-}
-export function payoutQuestId(record: { questId?: unknown; other?: unknown }): string {
-  const questId = typeof record.questId === "string" ? record.questId : "";
-  const other = typeof record.other === "string" ? record.other : "";
-  return questId || other.match(/QST-\d+/)?.[0] || "";
-}
-export function completedPayoutQuests(record: LegacyRecord): LegacyRecord[] {
-  return data.quests.filter(
-    (quest) =>
-      quest.status === "Completed" &&
-      (quest.selectedParticipant === record.title ||
-        quest.teamParticipants?.some(([name]) => name === record.title)),
-  );
 }
 function payoutQuestHistory(record: LegacyRecord): string {
   const quests = completedPayoutQuests(record);
@@ -894,15 +476,6 @@ function reportStatusTone(report: LegacyRecord): string {
   const label = reportStatusLabel(report);
   return label === "Open" ? "warning" : "neutral";
 }
-export function userReportsFor(user: LegacyRecord): LegacyRecord[] {
-  return (data.reports || [])
-    .filter((report) => report.reportedUserId === user.id)
-    .sort((first, second) => {
-      const parseReportedAt = (value: unknown) =>
-        Date.parse(String(value || "").replace(" · ", " ").replace(/\s+ICT$/, "")) || 0;
-      return parseReportedAt(second.reportedAt) - parseReportedAt(first.reportedAt);
-    });
-}
 function userReportCounts(reports: LegacyRecord[]): { open: number; closed: number } {
   return reports.reduce(
     (counts, report) => {
@@ -912,14 +485,6 @@ function userReportCounts(reports: LegacyRecord[]): { open: number; closed: numb
       return counts;
     },
     { open: 0, closed: 0 },
-  );
-}
-export function userQuestRecords(user: LegacyRecord): LegacyRecord[] {
-  return data.quests.filter(
-    (quest) =>
-      quest.person === user.title ||
-      quest.selectedParticipant === user.title ||
-      quest.teamParticipants?.some(([name]) => name === user.title),
   );
 }
 function userAccountSection(user: LegacyRecord): string {
@@ -985,9 +550,7 @@ function openUserReportDetails(user: LegacyRecord, report: LegacyRecord): void {
   drawer.querySelector<LegacyDomElement>("#close")?.addEventListener("click", closeDrawer);
   drawer.querySelector<LegacyDomElement>("#close-user-report")?.addEventListener("click", closeDrawer);
   drawer.querySelector<LegacyDomElement>("#back-to-user")?.addEventListener("click", () => {
-    const trigger = drawerTrigger;
     openDrawer("users", data.users.indexOf(user));
-    drawerTrigger = trigger;
   });
 }
 function openReportDrawer(index: number): void {
@@ -1092,7 +655,7 @@ export function openDrawer(v: string, i: number): void {
 }
 
 function openUserReportDialog(user: LegacyRecord): void {
-  activeCustomLayerClose?.();
+  closeActiveLayer();
   const reporter = data.users.find((candidate) => candidate.id !== user.id);
   if (!reporter) return;
   const overlay = document.createElement("div");
@@ -1152,9 +715,6 @@ function openUserReportDialog(user: LegacyRecord): void {
   });
 }
 
-export function addUserHistory(user: LegacyRecord, entry: LegacyHistoryEntry): void {
-  user.moderationHistory = [entry, ...(user.moderationHistory || [])];
-}
 function refreshUserAfterMutation(user: LegacyRecord): boolean {
   if (window.__KUQUEST_USER_DETAIL__?.user?.id === user.id) {
     window.__KUQUEST_USER_DETAIL__.render();
@@ -1164,104 +724,8 @@ function refreshUserAfterMutation(user: LegacyRecord): boolean {
   else if (state.view === "users") render();
   return false;
 }
-export const penaltyPolicy = Object.freeze({
-  redFlagDays: 7,
-  temporaryBanDays: 7,
-  newUserExemptionCount: 10,
-  postBanExemptionCount: 3,
-});
-export function confirmedViolationCount(user: LegacyRecord): number {
-  const fallback = user.status === "Red Flag" ? 1 : user.status === "Temp ban" ? 2 : user.status === "Perm ban" ? 3 : 0;
-  return Math.max(0, Number(user.confirmedViolationCount ?? fallback) || 0);
-}
-export function redFlagExemptionFor(user: LegacyRecord): PenaltyExemption | null {
-  const newUserRemaining = Math.max(0, Number(user.newUserExemptionRemaining) || 0);
-  if (newUserRemaining) return { field: "newUserExemptionRemaining", remaining: newUserRemaining, label: "new user (PC-12)" };
-  const postBanRemaining = Math.max(0, Number(user.postBanExemptionRemaining) || 0);
-  if (postBanRemaining) return { field: "postBanExemptionRemaining", remaining: postBanRemaining, label: "post-ban (PC-13)" };
-  return null;
-}
-export function penaltyOutcomeFor(user: LegacyRecord): PenaltyOutcome {
-  const violationNumber = confirmedViolationCount(user) + 1;
-  if (violationNumber === 1) {
-    return { key: "red-flag", label: "Red Flag", status: "Red Flag", tone: "warning", durationDays: penaltyPolicy.redFlagDays };
-  }
-  if (violationNumber === 2) {
-    return { key: "temporary-ban", label: "Temporary ban", status: "Temp ban", tone: "danger", durationDays: penaltyPolicy.temporaryBanDays };
-  }
-  return { key: "permanent-ban", label: "Permanent ban", status: "Perm ban", tone: "danger" };
-}
-export function penaltyOutcomeLabel(outcome: PenaltyOutcome | null): string {
-  if (!outcome) return "No penalty determined";
-  if (outcome.key === "red-flag-exempted") return "Red Flag exempted";
-  return `${outcome.label}${outcome.durationDays ? ` · ${outcome.durationDays} days` : ""}`;
-}
-function consumeRedFlagExemption(user: LegacyRecord, exemption: PenaltyExemption | null): void {
-  if (!exemption) return;
-  user[exemption.field] = Math.max(0, exemption.remaining - 1);
-}
-function recordConfirmedViolation(user: LegacyRecord, reason: string, note: string): PenaltyOutcome {
-  const previousStatus = user.status;
-  const appliedAt = adminDateTime();
-  const appliedBy = currentAdminName();
-  const violationNumber = confirmedViolationCount(user) + 1;
-  const nextOutcome = penaltyOutcomeFor(user);
-  const exemption = nextOutcome.key === "red-flag" ? redFlagExemptionFor(user) : null;
-  const outcome = exemption
-    ? { ...nextOutcome, key: "red-flag-exempted", status: "Normal", tone: "success", durationDays: null }
-    : nextOutcome;
-  user.confirmedViolationCount = violationNumber;
-  consumeRedFlagExemption(user, exemption);
-  let expiresAt = "";
-  if (outcome.durationDays) {
-    const expires = new Date();
-    expires.setDate(expires.getDate() + outcome.durationDays);
-    expiresAt = adminDateTime(expires);
-  }
-  const event = outcome.key === "red-flag-exempted"
-    ? "Violation recorded (Red Flag exempted)"
-    : `${outcome.label} applied`;
-  user.status = outcome.status;
-  user.tone = outcome.tone;
-  user.statusReason = reason;
-  user.statusAppliedAt = appliedAt;
-  user.statusAppliedBy = appliedBy;
-  delete user.redFlagExpiresAt;
-  delete user.banExpiresAt;
-  delete user.penalty;
-  if (outcome.key !== "red-flag-exempted") {
-    if (outcome.key === "red-flag") user.redFlagExpiresAt = expiresAt;
-    if (outcome.key === "temporary-ban") user.banExpiresAt = expiresAt;
-    user.penalty = {
-      label: outcome.label,
-      reason,
-      recordedAt: appliedAt,
-      appliedBy,
-      ...(outcome.durationDays ? { durationDays: outcome.durationDays } : {}),
-      ...(expiresAt ? { expiresAt } : {}),
-    };
-  }
-  user.age = outcome.key === "red-flag-exempted"
-    ? "Violation recorded · Red Flag exempted"
-    : expiresAt
-      ? `${outcome.status} · expires ${expiresAt}`
-      : outcome.status;
-  addUserHistory(user, {
-    event,
-    at: appliedAt,
-    by: appliedBy,
-    reason,
-    previousStatus,
-    newStatus: outcome.status,
-    violationNumber,
-    outcome: penaltyOutcomeLabel(outcome),
-  });
-  if (note) user.adminNotes = [{ at: appliedAt, by: appliedBy, note }, ...(user.adminNotes || [])];
-  refreshNavigationCounts();
-  return outcome;
-}
 export function openPenaltyDialog(user: LegacyRecord): void {
-  activeCustomLayerClose?.();
+  closeActiveLayer();
   const overlay = document.createElement("div");
   overlay.className = "party-chat-overlay";
   overlay.innerHTML = `<section class="party-chat-modal penalty-modal" role="dialog" aria-modal="true" aria-label="Confirm violation for ${escapeActivityText(user.title)}"><div class="chat-modal-head"><div><strong>Confirm violation</strong><small>${escapeActivityText(user.title)} · ${escapeActivityText(user.id)}</small></div><button class="icon close-party-chat" aria-label="Close penalty form"><span class="close-lines"></span></button></div><form class="penalty-form"><p class="chat-intro">Confirm that this account committed an actual policy violation. The SRS penalty ladder applies the next consequence automatically.</p><section class="penalty-policy-note" aria-label="Penalty ladder"><strong>Penalty ladder</strong><span>1st violation: Red Flag · 7 days</span><span>2nd violation: Temporary ban · 7 days</span><span>3rd violation: Permanent ban</span></section><section class="penalty-preview" data-penalty-preview aria-live="polite"></section><label for="penalty-reason">Reason for confirmed violation<textarea id="penalty-reason" name="reason" rows="3" minlength="8" maxlength="500" required aria-describedby="penalty-reason-help penalty-error" placeholder="State the evidence and policy behind this violation…"></textarea><small id="penalty-reason-help">Enter 8–500 characters explaining the evidence and policy.</small></label><label for="penalty-note">Internal admin note (optional)<textarea id="penalty-note" name="note" rows="2" maxlength="500" placeholder="Add context for authorized moderators…"></textarea></label><p class="login-error penalty-error" id="penalty-error" role="alert" hidden></p><button class="btn danger" type="submit">Confirm violation</button></form></section>`;
@@ -1294,6 +758,7 @@ export function openPenaltyDialog(user: LegacyRecord): void {
       return;
     }
     const outcome = recordConfirmedViolation(user, reason, form.elements.note.value.trim());
+    refreshNavigationCounts();
     persistAdminData();
     recordActivity(`Violation confirmed · ${penaltyOutcomeLabel(outcome)}`, `${user.id} · ${user.title} · ${reason}`);
     close();
@@ -1303,7 +768,7 @@ export function openPenaltyDialog(user: LegacyRecord): void {
   updateFields();
 }
 function openAdminNoteDialog(user: LegacyRecord): void {
-  activeCustomLayerClose?.();
+  closeActiveLayer();
   const overlay = document.createElement("div");
   overlay.className = "party-chat-overlay";
   overlay.innerHTML = `<section class="party-chat-modal penalty-modal" role="dialog" aria-modal="true" aria-label="Add admin note for ${escapeActivityText(user.title)}"><div class="chat-modal-head"><div><strong>Add admin note</strong><small>${escapeActivityText(user.title)} · ${escapeActivityText(user.id)}</small></div><button class="icon close-party-chat" aria-label="Close admin note form"><span class="close-lines"></span></button></div><form class="penalty-form admin-note-form"><p class="chat-intro">This note is visible only to authorized moderation staff.</p><label>Internal note<textarea name="note" rows="5" minlength="4" maxlength="500" required placeholder="Record useful moderation context…"></textarea></label><p class="login-error penalty-error" role="alert" hidden></p><button class="btn primary" type="submit">Save note</button></form></section>`;
@@ -1332,25 +797,6 @@ function openAdminNoteDialog(user: LegacyRecord): void {
     toast(`Admin note saved for ${user.title}.`);
   });
 }
-export function applyDemoAction(action: string, record: LegacyRecord): void {
-  const transitions: Record<string, [string, string]> = {
-    "Restrict user": ["Temp ban", "danger"],
-    "Set normal": ["Normal", "success"],
-    "Lift penalty": ["Normal", "success"],
-    "Hide quest": ["Hidden", "neutral"],
-    "Reject payout": ["Rejected", "danger"],
-    "Approve payout": ["Processing", "info"],
-    "Close report": ["Closed", "neutral"],
-    "Terminate quest": ["Cancelled", "cancelled"],
-  };
-  const next = transitions[action];
-  if (!next) return;
-  [record.status, record.tone] = next;
-  if (action === "Lift penalty") delete record.penalty;
-  record.age = "Just now";
-  if (action === "Close report") record.closedAt = reportDateTime();
-  refreshNavigationCounts();
-}
 function refreshNavigationCounts(): void {
   const counts = {
     disputes: data.disputes.filter((record) => record.status === "Active").length,
@@ -1362,115 +808,21 @@ function refreshNavigationCounts(): void {
     if (counter) counter.textContent = String(count);
   });
 }
-export function applyReportDecision(report: LegacyRecord, decision: string, reason: string): void {
-  const user = data.users.find((candidate) => candidate.id === report.reportedUserId),
-    resolvedAt = adminDateTime(),
-    resolvedBy = currentAdminName();
-  const outcome = decision === "confirmed-violation" && user
-    ? recordConfirmedViolation(user, reason, "")
-    : null;
-  report.status = "Closed";
-  report.tone = "neutral";
-  report.closedAt = reportDateTime();
-  report.decision = decision;
-  report.decisionLabel = decision === "no-violation"
-    ? "No violation"
-    : `Violation confirmed · ${penaltyOutcomeLabel(outcome)}`;
-  report.decisionDays = outcome?.durationDays || null;
-  report.decisionReason = reason;
-  report.resolution = decision === "no-violation"
-    ? "Report dismissed; no policy violation found."
-    : `${report.decisionLabel}.`;
-  report.resolvedBy = resolvedBy;
-  report.resolutionAt = resolvedAt;
-  if (!user) {
-    refreshNavigationCounts();
-    return;
-  }
-  if (decision === "no-violation") {
-    addUserHistory(user, { event: "Report dismissed", at: resolvedAt, by: resolvedBy, reason, previousStatus: user.status, newStatus: user.status });
-  }
-  refreshNavigationCounts();
-}
-function timelineDetail(title: string, index: number): string {
-  if (index === 0) return "Current state recorded in the audit trail";
-  if (/^Quest published/i.test(title)) {
-    return "The quest became available under its published terms";
-  }
-  if (/Applications received/i.test(title)) {
-    return "Applicants were recorded for review against the quest requirements";
-  }
-  if (/Applications closed/i.test(title)) {
-    return "The application window closed and no further applications were accepted";
-  }
-  if (/selected|accepted the terms/i.test(title)) {
-    return "Participant selection and acceptance were recorded for this quest";
-  }
-  if (/proposed terms change|change pending/i.test(title)) {
-    return "A proposed change is waiting for participant consent";
-  }
-  if (/submitted|evidence/i.test(title)) {
-    return "Submitted evidence was added to the record for review";
-  }
-  if (/field work started|work started/i.test(title)) {
-    return "The assigned participant began work on the quest";
-  }
-  if (/identity matched/i.test(title)) {
-    return "The account identity check was completed";
-  }
-  if (/account ownership verified/i.test(title)) {
-    return "The payout account was confirmed for the participant";
-  }
-  if (/quest funds reserved/i.test(title)) {
-    return "Quest funds were reserved for this payout";
-  }
-  if (/record created/i.test(title)) {
-    return "This record was added to the audit trail";
-  }
-  return "This event was recorded in the audit trail";
-}
-export function timeline(items: TimelineEntry[], options: TimelineOptions = {}): string {
-  return `<ul class="timeline">${items
-    .map((item, index) => {
-      const rawParts = typeof item === "string" ? String(item).split(" · ") : [];
-      const entry =
-        typeof item === "string"
-          ? {
-              title: rawParts.shift() || "Quest activity",
-              detail: rawParts.join(" · "),
-            }
-          : item,
-        parts = String(entry.detail || "").split(" · "),
-        timeParts = entry.time ? 0 : parts.length > 2 ? 2 : parts.length > 1 ? 1 : 0,
-        time = (entry.time || parts.slice(0, timeParts).join(" · ")).replace(/\s+ICT$/, ""),
-        detail = entry.time
-          ? entry.detail
-          : parts.slice(timeParts).join(" · ") || timelineDetail(entry.title, index);
-      return `<li><strong>${escapeActivityText(entry.title)}</strong>${time ? `<time>${escapeActivityText(time)}</time>` : ""}${options.showDetails === false || entry.showDetails === false ? "" : `<span>${escapeActivityText(detail)}</span>`}</li>`;
-    })
-    .join("")}</ul>`;
-}
-export function closeDrawer(): void {
-  if (drawerKeydown) {
-    drawer.removeEventListener("keydown", drawerKeydown);
-    drawerKeydown = null;
-  }
-  drawer.classList.remove("open");
-  drawer.setAttribute("aria-hidden", "true");
-  drawer.inert = true;
-  shell.inert = false;
-  const restore = drawerTrigger;
-  drawerTrigger = null;
-  setTimeout(() => {
-    scrim.hidden = true;
-    if (restore instanceof HTMLElement && restore.isConnected) restore.focus();
-  }, 220);
-}
 export function ensureDetailDrawer(view: string, index: number): void {
   if (view !== "quests" && view !== "disputes") return;
   const opener = view === "quests" ? "openQuestDrawer" : "openDisputeDrawer";
   const open = window[opener];
   if (open) open(index);
+}
+export function applyDemoAction(action: string, record: LegacyRecord): void {
+  if (applyDemoActionCore(action, record)) refreshNavigationCounts();
+}
+export function applyReportDecision(report: LegacyRecord, decision: string, reason: string): void {
+  applyReportDecisionCore(report, decision, reason);
+  refreshNavigationCounts();
+}
+export function autoRejectUnavailablePayout(record: LegacyRecord): boolean {
+  return autoRejectUnavailablePayoutCore(record);
 }
 const dialog = document.querySelector<LegacyDomElement>("#confirm");
 export function confirmAction(a: string, r: LegacyRecord, decisionDetail = "", onConfirm?: (reason: string) => void, options: ConfirmActionOptions = {}): void {
